For and while loops
Here I did the tutorial for for and while loops, as well as one of the
exercises from the link you sent me. I realized that I need to learn
more of this simple things. I think the best way is for every part of
the java learning to build a new project.